// src/pages/main-menu/payment/components/share-to-other/components/dialog/balanceShareRule/models/interfaces.ts
// Generated by https://quicktype.io

import { BalShareList } from "../../../../models/interfaces";

export interface BalanceShareDetailTable {
  detailEffDate?: string;
  detailExpDate?: string;
  acctResName?: string;
  balDesc?: string;
  priority?: number;
  operType?: string;
  balId?: number;
  acctResId?: number;
  subsId?: number;
}

// Generated by https://quicktype.io

export interface PayloadBALShareRule {
  acctId?: number;
  pSubsId?: number;
  routingId?: number;
  balShare: BALShare;
}

export interface BALShare {
  prefix?: string;
  accNbr?: string;
  ceilLimit?: number;
  dailyCeilLimit?: number;
  effDate?: string;
  expDate?: string;
  paymentForce?: string;
  acctId?: number;
  processType?: string;
  ownerSubsId?: number;
  balShareId?: number;
  // operType?: string;
  shareType?: string;
  children?: Child[];
}

export interface Child {
  balShareDetailId?: number;
  balShareId?: number;
  detailEffDate?: string;
  detailExpDate?: string;
  acctResName?: string;
  balDesc?: string;
  priority?: string;
  operType?: string;
  balId?: number;
  acctResId?: number;
  subsId?: number;
}

export interface mappingProps {
  proccesType: "ADD" | "EDIT" | "DELETE";
  item?: BalShareList;
}

export interface SubsInfoSimple {
  subsCatgNameList: null;
  subsSpecialGroupNameList: null;
  prodNextState: null;
  prodNextStateDate: null;
  servType: null;
  paidFlag: null;
  extMap: EXTMap;
  subsId: number;
  prefix: string;
  accNbr: string;
  custId: number;
  userId: number;
  agentId: null;
  orgId: number;
  orgName: null;
  areaId: number;
  areaName: null;
  updateDate: Date;
  routingId: number;
  defLangId: number;
  ppsPwd: null;
  comments: string;
  spId: number;
  acctId: number;
  acctNbr: null;
  prodId: number;
  offerId: number;
  offerName: null;
  completedDate: Date;
  activeDate: null;
  prodState: string;
  prodStateName: null;
  subsPlanId: number;
  packageFlag: string;
  parentProdId: null;
  indepProdId: null;
  prodStateDate: Date;
  prodUpdateDate: Date;
  createdDate: null;
  state: string;
  stateDate: Date;
  blockReason: string;
  blockReasonName: null;
  prodExpDate: null;
  agreementEffDate: null;
  agreementExpDate: null;
  prodAttrValueXML: null;
  subsPlanName: null;
  custName: null;
  certNbr: null;
  certTypeId: null;
  certTypeName: null;
  imsi: null;
}

export interface CheckNumber {
  subsInfo?: SubsInfoSimple;
  status: boolean;
}

export interface EXTMap {}